fix(project): fixing prisma generation.
