--- account/infra/database/prisma/schema.prisma
@@ -1,41 +1,0 @@
-datasource db {
-    provider = "postgres"
-    url      = env("DATABASE_URL")
-}
-
-// go run github.com/prisma/prisma-client-go db push --schema infra/database/schema.prisma
-// go run github.com/prisma/prisma-client-go migrate deploy --schema infra/database/schema.prisma
-
-generator db {
-    provider = "go run github.com/prisma/prisma-client-go"
-    output           = "./prisma"
-    package          = "prismaDB"
-}
-
-model Account {
-    id String @id @unique
-    email String @unique
-    firstname String
-    lastname String
-    admin Role @default(USER)
-    balance Int
-//    address Address?
-    createdAt DateTime @default(now())
-    updatedAt DateTime @updatedAt
-}
-
-model Address {
-    id              Int @id @default(autoincrement())
-//    account         Account @relation(fields: [accountID], references: [id])
-//    accountID       String
-    country         String
-    state           String
-    street          String
-    streetNumber    Int
-    complementary   String
-}
-
-enum Role {
-    USER
-    ADMIN
-}
