+++ account/infra/database/schema.prisma
@@ -1,6 +1,6 @@
 datasource db {
     provider = "postgres"
-    url      = env("PRISMA_URL")
+    url      = env("DATABASE_URL")
 }
 
 // go run github.com/prisma/prisma-client-go db push --schema infra/database/schema.prisma
@@ -8,7 +8,7 @@ datasource db {
 
 generator db {
     provider = "go run github.com/prisma/prisma-client-go"
-    output           = "./prisma"
+    output           = "prisma"
     package          = "prismaDB"
 }
